avoid ssize_t
git-svn-id: https://dev.csoft.net/agar/trunk@6468 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ core/load_xcf.c
@@ -327,7 +327,7 @@ xcf_read_tile_rle(AG_Netbuf *buf, Uint32 len, int bpp, int x, int y)
 {
 	int i, size, count, j;
 	Uint8 *tilep, *tile, *data;
-	ssize_t rv;
+	long rv;
 
 	tilep = tile = Malloc(len, M_LOADER);
 	rv = AG_NetbufReadE(tile, sizeof(Uint8), len, buf);
